Rename module so you can use 'ui' as variable name
