+++ src/euchre/interaction.py
